Extract logic for determining which eslint to use
Fixes #300

This allows the `Linter Eslint: Fix file` command to use the same ESLint
which is being used for linting the files, avoiding problems when plugins
or shared configs are used.

This also puts the execution of ESLint for error fixing onto the worker,
which will free up the UI from being blocked.  It also adds messages to
indicate the success or partial success of the operation.  ESLint will
report a 0 exit code only if all linting errors are removed.  Otherwise,
if any linting errors remain, it will exit with a 1, in which case a
warning will be displayed to indicate the operation is complete, but there
are still errors.

Also, this commit abstracts more methods into a helper, for reusability and testability.